refactor: member 도메인 구조 정리, GitHub api 요청하는 코드를 github-api 도메인으로 분리
