--- BE/src/common/types/githubResource.type.ts
@@ -1,4 +1,0 @@
-export interface GithubEmail {
-  primary: boolean;
-  email: string;
-}
