+++ BE/src/members/domain/entity/member.entity.ts
