+++ BE/src/members/members.module.ts
@@ -1,17 +1,21 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { MembersService } from './members.service';
-import { Member } from './entities/member.entity';
-import { MembersController } from './members.controller';
+import { MembersService } from './domain/service/members.service';
+import { Member } from './domain/entity/member.entity';
+import { MembersController } from './controller/members.controller';
 import { LesserJwtModule } from 'src/common/lesser-jwt/lesser-jwt.module';
 import { ConfigModule, ConfigService } from '@nestjs/config';
-import { MemberRepository } from './Repository/member.repository';
+import { MemberRepository } from './repository/member.repository';
+import { GithubOauthService } from 'src/github-api/oauth.service';
+import { GithubResourceService } from 'src/github-api/resource.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Member]), LesserJwtModule, ConfigModule.forRoot()],
   controllers: [MembersController],
   providers: [
     MembersService,
+    GithubOauthService,
+    GithubResourceService,
     ConfigService,
     {
       provide: 'MemberRepo',
