+++ BE/src/projects/projects.module.ts
@@ -8,7 +8,7 @@ import { AuthModule } from 'src/common/auth/auth.module';
 import { IsLoginGuard } from 'src/common/auth/IsLogin.guard';
 import { LesserJwtModule } from 'src/common/lesser-jwt/lesser-jwt.module';
 import { LesserJwtService } from 'src/common/lesser-jwt/lesser-jwt.service';
-import { Member } from 'src/members/entities/member.entity';
+import { Member } from 'src/members/domain/entity/member.entity';
 import { Sprint } from 'src/sprints/entities/sprint.entity';
 import { ProjectRepository } from './repository/Project.Repository';
 import { ProjectCounterRepository } from './repository/ProjectCounter.Repository';
