+++ BE/src/reviews/reviews.module.ts
@@ -7,7 +7,7 @@ import { ReviewsController } from './reviews.controller';
 import { Sprint } from 'src/sprints/entities/sprint.entity';
 import { Task } from 'src/backlogs/entities/task.entity';
 import { SprintToTask } from 'src/sprints/entities/sprint-task.entity';
-import { Member } from 'src/members/entities/member.entity';
+import { Member } from 'src/members/domain/entity/member.entity';
 import { Project } from 'src/projects/Domain/entity/project.entity';
 
 @Module({
